完成属性表与Map的联动，包括添加字段，编辑属性，以及对应要素的选择

+++ MalaSpiritGIS/MalaSpiritGIS/MLFeatures.cs
@@ -281,7 +281,7 @@ namespace MalaSpiritGIS
             else
             {
                 object[] curValues = new object[attributeData.Columns.Count];
-                curValues[0] = attributeData.Rows.Count;
+                curValues[0] = (uint)attributeData.Rows.Count;
                 curValues[1] = featureType;
                 attributeData.LoadDataRow(curValues, true);
             }
